% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRef_functions.R
\name{buildCellRef}
\alias{buildCellRef}
\title{Build a CellRef using the initial data integration and the seed cells}
\usage{
buildCellRef(
  object.all,
  object.seed,
  mapping.batch = "Dataset",
  integration.batch = "Dataset",
  integration.npcs = 200,
  rpca.prune_anchors = T,
  purity.pruning = T,
  purity.nn.k = 20,
  purity.thresh = 0.6,
  verbose = T
)
}
\arguments{
\item{object.all}{the Seurat object that contains all data for CellRef construction after batch correction}

\item{object.seed}{the Seurat object that contains the identified seed cells.}

\item{mapping.batch}{a character string indicating the meta.data column that contains data batch information to perform mapping to the seed cells}

\item{integration.npcs}{the number of principal components}
}
\value{
A Seurat object
}
\description{
Build a CellRef using the initial data integration and the seed cells
}
