% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRef_functions.R
\name{doDataIntegration}
\alias{doDataIntegration}
\title{Perform batch correction of single cell data}
\usage{
doDataIntegration(
  object,
  integration.batch,
  ident.var = NULL,
  method = c("Monocle3-mnn", "Seurat4-rpca", "Harmony"),
  npcs = 200,
  vars.to.regress = c("S.Score", "G2M.Score", "pMT"),
  umap.min_dist = 0.3,
  umap.n_neighbors = 50,
  do.clustering = TRUE,
  clustering.method = "leiden",
  clustering.resolution = NULL,
  rpca.normalization = "SCT",
  rpca.k.anchor = 10,
  rpca.k.weight = 50,
  rpca.prune_anchors = T,
  verbose = F,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{integration.batch}{a character string indicating the meta.data column that contains the batch for correction}

\item{ident.var}{a character string indicating the meta.data column that contains the identities of each cell}

\item{method}{the method used for batch correction}

\item{npcs}{the number of principal components used for batch correction and umap projection}

\item{vars.to.regression}{the covariates}
}
\value{
A Seurat object
}
\description{
This function loads a Seurat object that contains all data for batch correction and integration.
 It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
