% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRef_functions.R
\name{evaluatePrediction}
\alias{evaluatePrediction}
\title{CellRef validation of automated cell type annotation}
\usage{
evaluatePrediction(
  obj,
  markers,
  ident.use,
  umap.width = 10,
  umap.height = 7,
  log.thresh = log2(1.5),
  min.pct = 0.2,
  min.cells.group = 3,
  dotplot.height = 13,
  dotplot.width = 9,
  vlnplt.height = 5,
  vlnplt.width = 3,
  fullsigs.log.thrsh = log2(1.5),
  fullsigs.min.pct = 0.2,
  heatmap.height = 10,
  heatmap.width = 16,
  heatmap_n = 50,
  dotchart.height = 10,
  dotchart.width = 16,
  enrich.height = 6,
  enrich.width = 8
)
}
\arguments{
\item{obj}{A Seurat object}

\item{markers}{A data frame that contains marker(column name: 'Marker') and cell type(column name: 'CellType') information,}

\item{ident.use}{A character string indicating the meta.data column that contains the identities of each cell}

\item{umap.width}{Width of output UMAP}

\item{umap.height}{Height of output UMAP}

\item{log.thresh}{The threshold of Log Fold Change when finding the expression of CellRef marker genes for given ident.use}

\item{min.pct}{The min frequency to find the expression of CellRef marker genes for given ident.use}

\item{min.cells.group}{The smallest number of cells for a cell type in your annotation}

\item{dotplot.height}{Height of output dotplot}

\item{dotplot.width}{Width of output dotplot}

\item{vlnplt.height}{Height of output violin plot}

\item{fullsigs.log.thrsh}{The threshold of log.FC when finding full signature genes for given ident.use}

\item{fullsigs.min.pct}{The min frequency to find full signature genes for given ident}

\item{heatmap.height}{Height of output heatmap}

\item{heatmap.width}{Width of output heatmap}

\item{heatmap_n}{Number of DEGs per cell type in the heatmap}

\item{dotchart.height}{Height of dotchart}

\item{dotchart.width}{Width of dotchart}

\item{enrich.height}{Height of enrichment analysis plots}

\item{enrich.width}{Width of enrichment analysis plots}

\item{Vlnplt.width}{Width of output violin plot}
}
\value{
A PDF file and multiple data frames.
}
\description{
This function loads a Seurat object with cell type annotations predicted by CellRef or CellRef seed.
It will produce a PDF, containing UMAP visualization of cells colored by the CellRef predicted cell types,
differential expression of CellRef markers in the predictions, dotplot visualization of expression of CellRef marker genes,
identification of differentially expressed genes (DEGs) in the predicted cell types,
heatmap visualization of DEGs, and most enriched KEGG pathways and Gene Ontology Biological Processes enriched by the DEGs.
}
