% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRef_functions.R
\name{findSeedCells}
\alias{findSeedCells}
\title{Seed cell identification}
\usage{
findSeedCells(
  object,
  ctd,
  candidate.clusters,
  cluster.var = "integrated_clusters",
  exp.thresh = 0,
  use.scaled.exp = T,
  score.thresh = Inf,
  seed.n.min = 5,
  seed.n.max = 200,
  verbose = T
)
}
\arguments{
\item{object}{A Seurat object}

\item{ctd}{a data.frame object that contains the cell type dictionary for seed cell identification.}

\item{candidate.clusters}{a data.frame that contains the mapping between candidate cell clusters and cell types in the dictionary}

\item{cluster.var}{a character string indicating the meta.data column that contains cluster assignment of each cell}

\item{exp.thresh}{the number of principal components used for batch correction and umap projection}

\item{use.scaled.exp}{the covariates}
}
\value{
A Seurat object of the identified seed cells
}
\description{
Seed cell identification
}
