% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRef_functions.R
\name{buildCellRef}
\alias{buildCellRef}
\title{Build a CellRef using the initial data integration and the seed cells}
\usage{
buildCellRef(
  object.all,
  object.seed,
  mapping.batch = "Dataset",
  integration.batch = "Dataset",
  integration.npcs = 200,
  rpca.prune_anchors = T,
  purity.pruning = T,
  purity.nn.k = 20,
  purity.thresh = 0.6,
  verbose = T
)
}
\arguments{
\item{object.all}{A Seurat object that contains all data for CellRef construction after batch correction}

\item{object.seed}{The Seurat object that contains the identified seed cells. The object should have run UMAP project and have stored the UMAP model.}

\item{mapping.batch}{a character string indicating the meta.data column that contains data batch information to perform mapping to the seed cells}

\item{integration.batch}{the name of the meta.data column that contains the batch information of CellRef cells for correction using Seurat's RPCA method.}

\item{integration.npcs}{the number of principal components}

\item{rpca.prune_anchors}{In Seurat's RPCA integration, whether to prune anchors connecting cells of different identities.}

\item{purity.pruning}{Whether to prune cells based on kN purity}

\item{purity.nn.k}{The number of neighbors used in the kNN purity calculation}

\item{purity.thresh}{The threshold for kNN purity score}

\item{verbose}{Whether to show verbose output}
}
\value{
A Seurat object
}
\description{
Build a CellRef using the initial data integration and the seed cells
}
