% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRef_functions.R
\name{doDataIntegration}
\alias{doDataIntegration}
\title{Perform batch correction of single cell RNA-seq data}
\usage{
doDataIntegration(
  object,
  integration.batch,
  ident.var = NULL,
  method = c("Monocle3-mnn", "Seurat4-rpca", "Harmony"),
  npcs = 200,
  vars.to.regress = c("S.Score", "G2M.Score", "pMT"),
  umap.min_dist = 0.3,
  umap.n_neighbors = 50,
  do.clustering = TRUE,
  clustering.method = "leiden",
  clustering.resolution = NULL,
  rpca.normalization = "SCT",
  rpca.k.anchor = 10,
  rpca.k.weight = 50,
  rpca.prune_anchors = T,
  verbose = F,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{integration.batch}{a character string indicating the meta.data column that contains the batch for correction}

\item{ident.var}{a character string indicating the meta.data column that contains the identities of each cell}

\item{method}{the method used for batch correction}

\item{npcs}{the number of principal components used for batch correction and umap projection}

\item{umap.min_dist}{The effective minimum distance between embedded points, controlling how tightly the embedding is allowed to compress points together.}

\item{umap.n_neighbors}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation}

\item{do.clustering}{Whether to perform clustering analysis after batch correction}

\item{clustering.method}{The clustering method}

\item{clustering.resolution}{The resolution of graph-based clustering}

\item{rpca.normalization}{The normalization method used in Seurat's RPCA pipeline}

\item{rpca.k.anchor}{The k.anchor parameter used in Seurat's RPCA pipeline}

\item{rpca.k.weight}{The k.weight parameter used in Seurat's RPCA pipeline}

\item{rpca.prune_anchors}{When performing RPCA based integration, whether to prune anchors connecting cells of different identities. Need to set the ident.var parameter.}

\item{verbose}{Whether to show verbose output}

\item{vars.to.regression}{Variables to regress out. For example, cell cycle (S.Score, G2M.Score) and mitochondrial percentage (pMT).}
}
\value{
A Seurat object
}
\description{
This function loads a Seurat object that contains all data for batch correction and integration.
It will use the "RNA" assay to perform the data integration and assume that the "data" slot contains the normalized data.
Currently, the function supports three single cell batch correction methods
Monocle3-mnn: the mutual nearest neighbor matching based algorithm (Haghverdi et al., 2018) implemented in Monocle 3 (Cao et al., 2019)
Seurat4-rpca: the reciprocal PCA based integration in Seurat 4 (Hao and Hao, et al., 2021)
Harmony: the Harmony based integration (Korsunsky et al., 2019)
}
