% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRef_functions.R
\name{findCandidateClusters}
\alias{findCandidateClusters}
\title{Identify candidate cell clusters for each cell type defined in the dictionary}
\usage{
findCandidateClusters(
  object,
  ctd,
  cluster.var = "integrated_clusters",
  exp.thresh = 0,
  use.scaled.exp = T,
  score.thresh = 0.1,
  seed.n.min = 5,
  seed.n.max = Inf,
  group.size.min = 10,
  precision.thresh = 0.05,
  recall.thresh = 0.2,
  verbose = T
)
}
\arguments{
\item{object}{A Seurat object}

\item{ctd}{a data.frame object that contains the cell type dictionary}

\item{cluster.var}{a character string indicating the meta.data column that contains cluster assignment of each cell}
}
\value{
a data.frame containing the mapping of candidate cell clusters and cell types
}
\description{
Identify candidate cell clusters for each cell type defined in the dictionary
}
